feat: Prevent Signer from approving txns with crazy fees even when RBFing
