+++ signer/src/bitcoin/rpc.rs
@@ -23,12 +23,15 @@ use bitcoincore_rpc_json::GetNetworkInfoResult;
 use bitcoincore_rpc_json::GetRawTransactionResultVin;
 use bitcoincore_rpc_json::GetRawTransactionResultVout as BitcoinTxInfoVout;
 use bitcoincore_rpc_json::GetTxOutResult;
+use futures::future::try_join_all;
 use serde::Deserialize;
 use url::Url;
 
+use crate::bitcoin::utxo::Fees;
 use crate::bitcoin::BitcoinInteract;
 use crate::error::Error;
 
+use super::utxo;
 use super::GetTransactionFeeResult;
 use super::TransactionLookupHint;
 
@@ -730,4 +733,108 @@ impl BitcoinInteract for BitcoinCoreClient {
     async fn get_network_info(&self) -> Result<bitcoincore_rpc_json::GetNetworkInfoResult, Error> {
         self.get_network_info()
     }
+}
+
+/// Assesses the total fees paid for any outstanding sweep transactions in
+/// the mempool which may need to be RBF'd. If there are no sweep
+/// transactions which are spending the signer's UTXO, then this function
+/// will return [`None`].
+///
+/// TODO: This method currently blindly assumes that the mempool transactions
+/// are correct. Maybe we need some validation?
+#[tracing::instrument(skip_all, fields(signer_utxo = %signer_utxo.outpoint))]
+pub async fn assess_mempool_sweep_transaction_fees<BC>(
+    bitcoin_client: &BC,
+    signer_utxo: &utxo::SignerUtxo,
+) -> Result<Option<Fees>, Error>
+where
+    BC: BitcoinInteract + Clone + 'static,
+{
+    // Find the mempool transactions that are spending the provided UTXO.
+    let mempool_txs_spending_utxo = bitcoin_client
+        .find_mempool_transactions_spending_output(&signer_utxo.outpoint)
+        .await?;
+
+    // If no transactions are found, we have nothing to do.
+    if mempool_txs_spending_utxo.is_empty() {
+        tracing::debug!(
+            outpoint = %signer_utxo.outpoint,
+            "no mempool transactions found spending signer output; nothing to do"
+        );
+        return Ok(None);
+    }
+
+    tracing::debug!(
+        outpoint = %signer_utxo.outpoint,
+        "found mempool transactions spending signer output; assessing fees"
+    );
+
+    // If we have some transactions, we need to find the one that pays the
+    // highest fee. This is the transaction that we will use as the root of
+    // the sweep package. Note that even if only one transaction was
+    // returned above, we still need to get the fee for it, which is why
+    // there's no special logic for one vs multiple.
+    //
+    // This can technically error if the mempool transactions are not found,
+    // but it shouldn't happen since we got the transaction ids from
+    // bitcoin-core itself.
+    let best_sweep_root = try_join_all(mempool_txs_spending_utxo.iter().map(|txid| {
+        let bitcoin_client = bitcoin_client.clone();
+        async move {
+            bitcoin_client
+                .get_transaction_fee(txid, Some(TransactionLookupHint::Mempool))
+                .await
+                .map(|fee| (txid, fee))
+        }
+    }))
+    .await?
+    .into_iter()
+    .max_by_key(|(_, fees)| fees.fee);
+
+    // Since we got the transaction ids from bitcoin-core, these should
+    // not be missing, but we double-check here just in case (it could
+    // happen that the client has failed-over to the next node which isn't
+    // in sync with the previous one, for example).
+    let Some((best_sweep_root_txid, fees)) = best_sweep_root else {
+        tracing::warn!(
+            outpoint = %signer_utxo.outpoint,
+            "no fees found for mempool transactions spending signer output"
+        );
+        return Ok(None);
+    };
+
+    // Retrieve all descendant transactions of the best sweep root.
+    let descendant_txids = bitcoin_client
+        .find_mempool_descendants(best_sweep_root_txid)
+        .await?;
+
+    // Retrieve fees for all descendant transactions. If there were no
+    // descendants then this will just result in an empty list.
+    let descendant_fees = try_join_all(descendant_txids.iter().map(|txid| {
+        let bitcoin_client = bitcoin_client.clone();
+        async move {
+            bitcoin_client
+                .get_transaction_fee(txid, Some(TransactionLookupHint::Mempool))
+                .await
+        }
+    }))
+    .await?;
+
+    // Sum the fees of the best sweep root and its descendants, while also
+    // summing the vsize of the transactions for fee-rate calculation later.
+    // If there were no descendants then this will just be the fee and size
+    // from the best root sweep transaction.
+    let (total_fees, total_vsize) = descendant_fees
+        .into_iter()
+        .fold((fees.fee, fees.vsize), |acc, fees| {
+            (acc.0 + fees.fee, acc.1 + fees.vsize)
+        });
+
+    // Calculate the fee rate based on the total fees and vsizes of the
+    // transactions which we've found. Since this is returning transactions
+    // from bitcoin-core, we should have valid fees and sizes, so we don't
+    // need to check for division by zero.
+    let rate = total_fees as f64 / total_vsize as f64;
+
+    Ok(Some(Fees { total: total_fees, rate }))
 }
